+++ R/matchit2cardinality.R
@@ -73,7 +73,9 @@
 #'   free trial or academic license. 
 #'   \item `time`: the maximum amount of
 #'   time before the optimization routine aborts, in seconds. Default is 120 (2
-#'   minutes). For large problems, this should be set much higher.  
+#'   minutes). For large problems, this should be set much higher. When the limit is
+#'   reached, the best solution found so far is returned with a warning; see
+#'   *Dealing with Errors and Warnings* in Details.
 #'   }
 #'
 #'   The arguments `distance` (and related arguments), `replace`, `m.order`, and `caliper` (and related arguments) are ignored with a warning.
@@ -178,6 +180,12 @@
 #' size, the optimizers will stall at one of them, not thinking it has found
 #' the optimum. The result should be checked to see if it can be used as the
 #' solution.
+#'
+#' An error that says `"The optimizer failed to find any solution in the time allotted."`
+#' means that the time limit was reached before any solution satisfying the constraints
+#' was found, so there is nothing to return. Increasing `time` may allow one to be found,
+#' though the problem may also be infeasible (see below). Only HiGHS distinguishes this
+#' case; with GLPK and Gurobi, it produces the error about infeasibility.
 #'
 #' An error that says `"The optimization problem may be infeasible."`
 #' usually means that there is a issue with the optimization problem, i.e.,
@@ -715,7 +723,14 @@ cardinality_error_report <- function(out, solver) {
       arg::err("the optimization problem may be infeasible. Try increasing the value of {.arg tols}. See {.topic MatchIt::method_cardinality} for additional details")
     }
     if (out$status_message %in% c("Time limit reached", "Iteration limit reached")) {
-      arg::err("the optimizer failed to find an optimal solution in the time allotted. Try increasing the value of {.arg time}. See {.topic MatchIt::method_cardinality} for additional details")
+      #HiGHS keeps the best feasible solution found before the limit (the incumbent);
+      #when there is none, `primal_solution` is all zeros and must not be returned.
+      if (identical(out$info$primal_solution_status, "Feasible")) {
+        arg::wrn("the optimizer failed to find an optimal solution in the time allotted. The returned solution may not be optimal. See {.topic MatchIt::method_cardinality} for additional details")
+      }
+      else {
+        arg::err("the optimizer failed to find any solution in the time allotted. Try increasing the value of {.arg time}. See {.topic MatchIt::method_cardinality} for additional details")
+      }
     }
   }
 }